Refactor: Add new magazine to MagazineSlider component and update schema
Refactor the MagazineSlider component to include a new magazine entry for "Spark 8th Edition". Also, update the schema in the sanity folder to include the "content" field in the nmbDescription document type.

Refactor: Update NMBFocus component to render content from PortableText

Update the NMBFocus component to render the content using the PortableText component from the @portabletext/react library. This allows for more flexible and dynamic rendering of the description text.

Refactor: Update Hero component layout and content

Update the Hero component to improve the layout and content. This includes adjusting the positioning of the text and image, as well as updating the text content for the homepageHero subtitle.

Refactor: Update page.tsx to fetch and use content from NmbPageDescription

Update the page.tsx file to fetch and use the "content" field from the NmbPageDescription document in the getNmbPageDescription function. This allows for rendering of rich text content in the NMBFocus component.

Refactor: Update GlobalPageDescription component to include an ID for description

Update the GlobalPageDescription component to include an ID field for the description. This helps with identifying and referencing specific descriptions in the codebase.